Fix a potential issue in calling Cpu->GetTimerValue (Cpu, 0, &CurrentTick, (UINT64 *) &TimerPeriod);
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@11076 6f19259b-4bc3-4df7-8a09-765794883524

+++ NetworkPkg/Application/Ping6/Ping6.h
@@ -65,9 +65,9 @@ typedef struct _PING6_PRIVATE_DATA {
   EFI_IP6_COMPLETION_TOKEN    RxToken;
   UINT16                      RxCount;
   UINT16                      TxCount;
-  UINT32                      RttSum;
-  UINT32                      RttMin;
-  UINT32                      RttMax;
+  UINT64                      RttSum;
+  UINT64                      RttMin;
+  UINT64                      RttMax;
   UINT32                      SequenceNum;
 
   EFI_IPv6_ADDRESS            SrcAddress;
